+++ app/controllers/posts_controller.rb
@@ -5,6 +5,7 @@ class PostsController < ApplicationController
   end
 
   def create
+    
   end
 
   def index
