+++ controllers/static_pages_controller_test.rb
