--- test/controllers/posts_controller_test.rb
@@ -1,19 +1,0 @@
-require 'test_helper'
-
-class PostsControllerTest < ActionDispatch::IntegrationTest
-  test "should get new" do
-    get posts_new_url
-    assert_response :success
-  end
-
-  test "should get create" do
-    get posts_create_url
-    assert_response :success
-  end
-
-  test "should get index" do
-    get posts_index_url
-    assert_response :success
-  end
-
-end
