+++ test/models/user_test.rb
@@ -1,7 +1,13 @@
 require 'test_helper'
 
 class UserTest < ActiveSupport::TestCase
-  # test "the truth" do
-  #   assert true
-  # end
+  def setup
+    @user = users(:ivan)
+    @post = Post.new(user_id: @user.id)
+  end
+
+  test "Post should be valid" do
+    assert @post.valid?
+  end
+
 end
